backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
replicate
Pillow
aiofiles
python-multipart

